refactor: isolate llmfs generation lifecycle
